requirements.txt
scipy==1.8.1
numpy
mypy
h5py
matplotlib-stubs
pytest